Fixed several issues
The issue with the edit feature not working is now fixed. The problem was that the only thing keeping track of the changes to the cow was state. Because of this the program did not know what the existing values of the cow were and just set them to empty strings because that is what was being passed. Now, the current cow is extracted from the Card component when the edit button is clicked with the use of the find() function. It is then passed to the parent function and on to the editCattle component. The actual data sent to the api is now determined bythe currentCow object itself instead of the state variables. If the value of the state variables are not equal to empty strings only then can the value be set to the currentCow object and passed to the api. Also, an issue was fixed with getHerd and getCattle infinitly being called and killing performance with a simple modification to the useEffect hook.